--- sip-core/src/main/java/de/ikor/sip/foundation/core/framework/scope/exchange/notifiers/CompletedExchangeEventNotifier.java
@@ -1,24 +1,0 @@
-package de.ikor.sip.foundation.core.framework.scope.exchange.notifiers;
-
-import de.ikor.sip.foundation.core.framework.scope.exchange.ExchangeContextHolder;
-import org.apache.camel.impl.event.ExchangeCompletedEvent;
-import org.apache.camel.impl.event.ExchangeFailedEvent;
-import org.apache.camel.spi.CamelEvent;
-import org.apache.camel.support.EventNotifierSupport;
-
-public class CompletedExchangeEventNotifier extends EventNotifierSupport {
-
-  public void notify(CamelEvent event) {
-    resetContextHolderInstance();
-  }
-
-  private void resetContextHolderInstance() {
-    ExchangeContextHolder conversationContextHolder = ExchangeContextHolder.instance();
-    conversationContextHolder.resetConversationAttributes();
-  }
-
-  @Override
-  public boolean isEnabled(CamelEvent event) {
-    return event instanceof ExchangeCompletedEvent || event instanceof ExchangeFailedEvent;
-  }
-}
