+++ sip-core/src/main/java/de/ikor/sip/foundation/core/scope/CustomScopeConfig.java
@@ -1,10 +1,9 @@
-package de.ikor.sip.foundation.core.framework.scope;
+package de.ikor.sip.foundation.core.scope;
 
-import de.ikor.sip.foundation.core.framework.scope.conversation.ConversationScope;
+import de.ikor.sip.foundation.core.scope.conversation.ConversationScope;
+import de.ikor.sip.foundation.core.scope.exchange.ExchangeScope;
 import java.util.HashMap;
 import java.util.Map;
-
-import de.ikor.sip.foundation.core.framework.scope.exchange.ExchangeScope;
 import org.springframework.beans.factory.config.CustomScopeConfigurer;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
