+++ sip-core/src/main/java/de/ikor/sip/foundation/core/scope/ScopeAutoConfig.java
@@ -1,0 +1,6 @@
+package de.ikor.sip.foundation.core.scope;
+
+import org.springframework.context.annotation.ComponentScan;
+
+@ComponentScan
+public class ScopeAutoConfig {}
